Add add

+++ app/controllers/users/chats_controller.rb
@@ -17,8 +17,12 @@ class Users::ChatsController < ApplicationController
   def destroy
     @chat_room = ChatRoom.find(params[:chat_room_id])
     chat = @chat_room.chats.find(params[:id])
-    chat.destroy
-    redirect_to chat_room_path(@chat_room.id)
+    if chat.user_id != current_user.id
+      redirect_to chat_room_path(@chat_room.id)
+    else
+      chat.destroy
+      redirect_to chat_room_path(@chat_room.id)
+    end
   end
 
   def chat_params
